Fix strand mod color everywhere
