--- src/app/loadout/loadout-ui/PlugDef.m.scss
@@ -1,5 +1,0 @@
-.strandColorFix {
-  img {
-    filter: hue-rotate(255deg) brightness(1.3);
-  }
-}
