+++ src/app/loadout/loadout-ui/PlugDef.tsx
@@ -5,8 +5,6 @@ import { PluggableInventoryItemDefinition } from 'app/inventory/item-types';
 import { PlugDefTooltip } from 'app/item-popup/PlugTooltip';
 import { DestinyClass } from 'bungie-api-ts/destiny2';
 import clsx from 'clsx';
-import { PlugCategoryHashes } from 'data/d2/generated-enums';
-import styles from './PlugDef.m.scss';
 
 interface Props {
   plug: PluggableInventoryItemDefinition;
@@ -16,15 +14,6 @@ interface Props {
   automaticallyPicked?: boolean;
   forClassType: DestinyClass | undefined;
 }
-
-const strandWrongColorPlugCategoryHashes = [
-  PlugCategoryHashes.TitanStrandClassAbilities,
-  PlugCategoryHashes.HunterStrandClassAbilities,
-  PlugCategoryHashes.WarlockStrandClassAbilities,
-  PlugCategoryHashes.TitanStrandMovement,
-  PlugCategoryHashes.HunterStrandMovement,
-  PlugCategoryHashes.WarlockStrandMovement,
-];
 
 /**
  * Displays a plug (mod, perk) based on just its definition, with optional close button.
@@ -37,13 +26,10 @@ export default function PlugDef({
   automaticallyPicked,
   forClassType,
 }: Props) {
-  const needsStrandColorFix = strandWrongColorPlugCategoryHashes.includes(
-    plug.plug.plugCategoryHash
-  );
   const contents = (
     <div
       role={onClick ? 'button' : undefined}
-      className={clsx('item', className, { [styles.strandColorFix]: needsStrandColorFix })}
+      className={clsx('item', className)}
       onClick={onClick}
       // eslint-disable-next-line jsx-a11y/no-noninteractive-tabindex
       tabIndex={onClick ? 0 : undefined}
